Use pyAudio to play audio cue

+++ src/cue/volume.py
@@ -1,29 +1,29 @@
 class Volume:
-    def __init__(self, left: float = 0.0, right: float = 0.0) -> None:
+    def __init__(self, left: int = 0, right: int = 0) -> None:
         self.__left = left
         self.__right = right
 
     @property
-    def left(self) -> float:
+    def left(self) -> int:
         return self.__left
 
     @left.setter
-    def left(self, value: float) -> None:
-        if value < 0.0:
-            value = 0.0
+    def left(self, value: int) -> None:
+        if value < 0:
+            value = 0
         self.__left = value
 
     @property
-    def right(self) -> float:
+    def right(self) -> int:
         return self.__right
 
     @right.setter
-    def right(self, value: float) -> None:
-        if value < 0.0:
-            value = 0.0
+    def right(self, value: int) -> None:
+        if value < 0:
+            value = 0
         self.__right = value
 
-    def getVolume(self) -> tuple[float, float]:
+    def getVolume(self) -> tuple[int, int]:
         return (self.left, self.right)
 
     def __str__(self) -> str:
